Extracted ApplyAttribute class to IHP.HSX.Attribute
Fixes https://github.com/digitallyinduced/ihp/issues/1731
